requirements.txt
biopython>=1.80
pandas>=1.5.0
openpyxl>=3.0.0